requirements.txt
streamlit
mistralai
pinecone-client
sentence-transformers
langchain
pillow
google-generativeai
